# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para operaciones frecuentes de docentes y estudiantes en EduGo",
        "title": "EduGo API Mobile",
        "termsOfService": "http://edugo.com/terms/",
        "contact": {
            "name": "Equipo EduGo",
            "email": "soporte@edugo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Genera JWT token tras validar credenciales",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar usuario",
                "parameters": [
                    {
                        "description": "Credenciales de usuario",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/materials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene lista de materiales filtrados por unidad y materia con progreso del estudiante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener lista de materiales",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de unidad académica",
                        "name": "unit_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID de materia",
                        "name": "subject_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "all",
                            "new",
                            "in_progress",
                            "completed"
                        ],
                        "type": "string",
                        "description": "Estado del material",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MaterialListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo material educativo y retorna URL firmada para subir PDF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Crear nuevo material",
                "parameters": [
                    {
                        "description": "Datos del material",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/CreateMaterialResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene información completa de un material incluyendo URL firmada para descargar PDF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener detalle de material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MaterialDetailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}/assessment": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene las preguntas del quiz SIN respuestas correctas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener cuestionario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AssessmentResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}/assessment/attempts": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Valida respuestas del quiz y registra puntaje del estudiante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Registrar intento de evaluación",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Respuestas del estudiante",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RecordAttemptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AttemptResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/materials/{id}/upload-complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra versión del material y publica evento para procesamiento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Notificar que el upload se completó",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del archivo subido",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UploadCompleteRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/UploadCompleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Answer": {
            "type": "object",
            "required": [
                "answer",
                "question_id"
            ],
            "properties": {
                "answer": {
                    "type": "string",
                    "example": "option_a"
                },
                "question_id": {
                    "type": "string",
                    "example": "q-uuid-1"
                }
            }
        },
        "AssessmentResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Evaluación de conceptos básicos"
                },
                "passing_score": {
                    "type": "integer",
                    "example": 70
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Question"
                    }
                },
                "time_limit_minutes": {
                    "type": "integer",
                    "example": 15
                },
                "title": {
                    "type": "string",
                    "example": "Cuestionario: Introducción a Pascal"
                },
                "total_points": {
                    "type": "integer",
                    "example": 100
                },
                "total_questions": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "AttemptResultResponse": {
            "type": "object",
            "properties": {
                "detailed_feedback": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/QuestionFeedback"
                    }
                },
                "passed": {
                    "type": "boolean",
                    "example": true
                },
                "score": {
                    "type": "number",
                    "example": 85.5
                },
                "total_points": {
                    "type": "number",
                    "example": 100
                }
            }
        },
        "CreateMaterialRequest": {
            "type": "object",
            "required": [
                "description",
                "file_name",
                "file_size",
                "subject_id",
                "title",
                "unit_id"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Material base sobre programación en Pascal"
                },
                "file_name": {
                    "type": "string",
                    "example": "introduccion-pascal.pdf"
                },
                "file_size": {
                    "type": "integer",
                    "example": 2048576
                },
                "subject_id": {
                    "type": "string",
                    "example": "subject-uuid-123"
                },
                "title": {
                    "type": "string",
                    "example": "Introducción a Pascal"
                },
                "unit_id": {
                    "type": "string",
                    "example": "unit-uuid-456"
                }
            }
        },
        "CreateMaterialResponse": {
            "type": "object",
            "properties": {
                "material_id": {
                    "type": "string",
                    "example": "uuid-123"
                },
                "max_file_size_bytes": {
                    "type": "integer",
                    "example": 104857600
                },
                "status": {
                    "type": "string",
                    "example": "created"
                },
                "upload_url": {
                    "type": "string",
                    "example": "https://s3.amazonaws.com/edugo-materials-prod/upload-url-mock"
                },
                "upload_url_expires_at": {
                    "type": "string",
                    "example": "2025-01-29T12:15:00Z"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "Field 'title' is required"
                }
            }
        },
...
        "LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "maria@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "LoginResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2025-01-29T12:00:00Z"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/UserInfo"
                }
            }
        },
        "MaterialDetail": {
            "type": "object",
            "properties": {
                "author_name": {
                    "type": "string",
                    "example": "Prof. García"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-01-15T12:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Material base sobre Pascal"
                },
                "file_size": {
                    "type": "integer",
                    "example": 2048576
                },
                "id": {
                    "type": "string",
                    "example": "uuid-1"
                },
                "my_progress": {
                    "type": "number",
                    "example": 0
                },
                "subject_name": {
                    "type": "string",
                    "example": "Programación"
                },
                "title": {
                    "type": "string",
                    "example": "Introducción a Pascal"
                }
            }
        },
        "MaterialDetailResponse": {
            "type": "object",
            "properties": {
                "has_quiz": {
                    "type": "boolean",
                    "example": true
                },
                "has_summary": {
                    "type": "boolean",
                    "example": true
                },
                "material": {
                    "$ref": "#/definitions/MaterialDetail"
                },
                "pdf_url": {
                    "type": "string",
                    "example": "https://s3.../presigned-url"
                },
                "pdf_url_expires_at": {
                    "type": "string",
                    "example": "2025-01-29T11:15:00Z"
                }
            }
        },
        "MaterialListResponse": {
            "type": "object",
            "properties": {
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MaterialSummary"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "MaterialSummary": {
            "type": "object",
            "properties": {
                "has_quiz": {
                    "type": "boolean",
                    "example": true
                },
                "has_summary": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "string",
                    "example": "uuid-1"
                },
                "progress": {
                    "type": "number",
                    "example": 0
                },
                "published_at": {
                    "type": "string",
                    "example": "2025-01-15T12:00:00Z"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "new",
                        "in_progress",
                        "completed"
                    ],
                    "example": "new"
                },
                "subject_name": {
                    "type": "string",
                    "example": "Programación"
                },
                "title": {
                    "type": "string",
                    "example": "Introducción a Pascal"
                },
                "unit_name": {
                    "type": "string",
                    "example": "5.º A - Programación"
                }
            }
        },
...
        "RecordAttemptRequest": {
            "type": "object",
            "required": [
                "answers"
            ],
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Answer"
                    }
                }
            }
        },
...
        "UploadCompleteRequest": {
            "type": "object",
            "required": [
                "checksum",
                "file_size",
                "version_key"
            ],
            "properties": {
                "checksum": {
                    "type": "string",
                    "example": "sha256:abc123..."
                },
                "file_size": {
                    "type": "integer",
                    "example": 2048576
                },
                "version_key": {
                    "type": "string",
                    "example": "materials/uuid-1/v1.pdf"
                }
            }
        },
        "UploadCompleteResponse": {
            "type": "object",
            "properties": {
                "material_id": {
                    "type": "string",
                    "example": "uuid-123"
                },
                "message": {
                    "type": "string",
                    "example": "Material queued for processing"
                },
                "status": {
                    "type": "string",
                    "example": "processing"
                }
            }
        },
        "UserInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "maria@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "user-uuid-123"
                },
                "name": {
                    "type": "string",
                    "example": "María González"
                },
                "role": {
                    "type": "string",
                    "example": "teacher"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token con formato: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}